requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
requests>=2.31.0
duckduckgo-search>=6.1.0
jinja2>=3.1.0
python-dateutil>=2.8.0 